7624 Konverter operations.jsx til TypeScript og refaktorer typer
 - Konverter operations.jsx til operations.ts med komplette TypeScript-typer
 - Legg til typer på alle funksjoner (hent, resetJournalforing, prepareKnyttTilSakForm)
 - Flytt Sak og FeltNavn interfaces fra knyttTilSak.tsx til ducks/journalforing/types.ts
 - Fjern circular dependency mellom operations og UI-komponent
 - Forbedre feilhåndtering: begrens .catch() til kun API-kall, ikke filtreringslogikk
